add get_degree staticmethod to PNAConv
